lab3/requirements.txt
tensorflow>=2.0.0
keras>=2.3.0
keras-vggface
mtcnn
numpy
matplotlib
Pillow
wget
opencv-python
picamera
python-telegram-bot
